pyproject.toml
[project]
name = "gemini-experimental-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.2.0",
    "pyttsx3>=2.98",
    "requests>=2.32.3",
]
